added finalized readme and fixed issues with some of the functions

+++ functions/remove.js
@@ -11,6 +11,7 @@ const removeEmployee = () => {
         const empArray = rows.map((employees) => {
             return employees.first_name;
         });
+        empArray.push('X None')
 
         inquirer.prompt(
             {
@@ -21,24 +22,28 @@ const removeEmployee = () => {
             }
         ).then((answer) => {
 
-            //finding the employees id
-            sql = "SELECT id FROM employees WHERE first_name='" + answer.empRemove + "'";
-            connection.query(sql, (err, value) => {
-                if (err) throw ('no employee found with that value');
-                
-                //finding where that employees id falls in the manager role and deleting them from their
-                sql = "UPDATE employees SET manager_id=null WHERE manager_id="+value[0].id;
-                connection.query(sql, (err) => {
-                    if (err) throw err;
-                });
-            });
+            if (answer.empRemove == 'X None') {
+                return;
+            } else {
+                //finding the employees id
+                sql = "SELECT id FROM employees WHERE first_name='" + answer.empRemove + "'";
+                connection.query(sql, (err, value) => {
+                    if (err) throw ('no employee found with that value');
 
-            //fully deleting that employee form the database
-            sql = "DELETE FROM employees WHERE first_name='" + answer.empRemove + "'";
-            connection.query(sql, (err, row) => {
-                if (err) throw err;
-                console.log("Employee "+answer.empRemove+" removed from the database!")
-            });
+                    //finding where that employees id falls in the manager role and deleting them from their
+                    sql = "UPDATE employees SET manager_id=null WHERE manager_id=" + value[0].id;
+                    connection.query(sql, (err) => {
+                        if (err) throw err;
+                    });
+                });
+
+                //fully deleting that employee form the database
+                sql = "DELETE FROM employees WHERE first_name='" + answer.empRemove + "'";
+                connection.query(sql, (err, row) => {
+                    if (err) throw err;
+                    console.log("Employee " + answer.empRemove + " removed from the database!")
+                });
+            }
         });
     });
 };
@@ -61,7 +66,7 @@ const editEmployee = () => {
                 return role.title;
             });
 
-            inquirer.prompt(
+            inquirer.prompt([
                 {
                     type: "list",
                     name: "empToEdit",
@@ -72,19 +77,19 @@ const editEmployee = () => {
                     type: "list",
                     name: "empEditChoice",
                     message: "What would you like to edit about the employee?",
-                    choices: ['first_name', 'last_name', 'role', 'manager']
+                    choices: ['first name', 'last name', 'role', 'manager']
                 },
                 {
                     type: "input",
                     name: "empEditFirstName",
                     message: "What is this employees new First Name?",
-                    when: (answers) => answers.empEditChoice == 'first_name',
+                    when: (answers) => answers.empEditChoice == 'first name',
                 },
                 {
                     type: "input",
                     name: "empEditLastName",
                     message: "What is this employees new Last Name?",
-                    when: (answers) => answers.empEditChoice == 'last_name',
+                    when: (answers) => answers.empEditChoice == 'last name',
                 },
                 {
                     type: "list",
@@ -100,10 +105,10 @@ const editEmployee = () => {
                     choices: empArray,
                     when: (answers) => answers.empEditChoice == 'manager',
                 }
-            ).then((answers) => {
+            ]).then((answers) => {
 
                 //return the id of the selected employee
-                let sql = "SELECT id FROM employees WHERE first_name='" + answers.empRemove + "'";
+                let sql = "SELECT id FROM employees WHERE first_name='" + answers.empToEdit + "'";
                 connection.query(sql, (err, id) => {
                     if (err) throw err;
 
@@ -111,26 +116,32 @@ const editEmployee = () => {
                     switch (answers.empEditChoice) {
 
                         //chose to edit first_name option running query to input new name into chosen employee field
-                        case "first_name":
-                            let sql = "UPDATE employees SET first_name = '" + answers.empEditFirstName + "' WHERE employees.id=" + id[0].id;
+                        case "first name":
+                            sql = "UPDATE employees SET first_name = '" + answers.empEditFirstName + "' WHERE id=" + id[0].id;
+                            console.log(sql)
                             connection.query(sql, (err) => {
                                 if (err) throw err;
                             });
                             break;
 
                         //chose to edit last_name option running query to input new last name into chosen emp field
-                        case "last_name":
-                            sql = "UPDATE employees SET last_name = '" + answers.empEditLastName + "' WHERE employees.id=" + id;
+                        case "last name":
+                            sql = "UPDATE employees SET last_name = '" + answers.empEditLastName + "' WHERE id=" + id[0].id;
                             connection.query(sql, (err) => {
                                 if (err) throw err;
                             });
                             break;
-                            
+
                         //chose to edit role option running query to input new role into chosen emp field
                         case "role":
-                            sql = "UPDATE role SET title='" + answers.empEditRole + "' WHERE employees.id=" + id;
-                            connection.query(sql, (err) => {
+                            sql = "SELECT id FROM role WHERE title='" + answers.empEditRole + "'";
+                            connection.query(sql, (err, roleId) => {
+
                                 if (err) throw err;
+                                sql = "UPDATE employees SET role_id='" + roleId[0].id + "' WHERE id=" + id[0].id;
+                                connection.query(sql, (err) => {
+                                    if (err) throw err;
+                                });
                             });
                             break;
 
@@ -138,19 +149,19 @@ const editEmployee = () => {
                         case "manager":
 
                             //finding id of manager from first name that was selected
-                            sql = "SELECT id FROM employees WHERE first_name=" + answers.empEditManager;
+                            sql = "SELECT id FROM employees WHERE first_name='" + answers.empEditManager + "'";
                             connection.query(sql, (err, managerId) => {
                                 if (err) throw err;
 
                                 //updating that employees manager to the one selected
-                                sql = "UPDATE employees SET manager_id=" + managerId + "WHERE employees.id=" + id;
-                                connection.query(sql, (err, managerId) => {
+                                sql = "UPDATE employees SET manager_id=" + managerId[0].id + " WHERE employees.id=" + id[0].id;
+                                connection.query(sql, (err) => {
                                     if (err) throw err;
                                 });
                             });
                             break;
                     }
-                    console.log("Employee "+ answers.empEditFirstName +" "+ answers.empEditLastName +" Successfully edited!")
+                    console.log("Employee " + answers.empToEdit + " Successfully edited!")
                     //end switch statement
                 });
             });
